20250315_3
Finally got a spawned building to function correctly with its correct script, you have to pass/set all of the resource variables to a new variable within the script that you want to use them "instantiate" of a resouces doesn't populate it with a variable that is each of the items in its consistent array of data.  Trash pile is broken for some reason and will no longer spawn.  Try duplicating the found mound, and or the tread bundle to correct that specific building scene