20250706
Got the signal for clicking on a specific drop and giving its name to the raid menu for the next round of data sourcing to add to the "selected raid drops" part of the raid menu. also added a resource and resouce group for drops and created a food drop to start to setup randomization of the drops, setting the texture was the next item, that is not yet finished.